refactor: extract vote types/choices to its own component
